requirements.txt
aiomysql==0.2.0
aiopg==1.4.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.11.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncmy==0.2.9
asyncpg==0.29.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
crypto==1.4.1
databases==0.9.0
decorator==5.1.1
Django==5.1.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
Naked==0.1.32
numerize==0.12
numpy==2.1.3
orjson==3.10.6
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.37.2
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
